refactor(contentful-client): restructure into management/delivery/helpers modules
